♻️ :: Operate Event
