+++ presentation/src/main/java/com/example/dms_android/feature/auth/login/SignInEvent.kt
@@ -5,11 +5,4 @@ import com.example.dms_android.base.MviEvent
 sealed class SignInEvent : MviEvent {
     data class InputId(val id: String) : SignInEvent()
     data class InputPassword(val password: String) : SignInEvent()
-    object SignInSuccess : SignInEvent()
-    object BadRequestException : SignInEvent()
-    object UnAuthorizedException : SignInEvent()
-    object NotFoundException : SignInEvent()
-    object TooManyRequestException : SignInEvent()
-    object InternalServerException : SignInEvent()
-    object UnKnownException : SignInEvent()
 }
