+++ presentation/src/main/java/com/example/dms_android/feature/auth/login/SignInViewEffect.kt
@@ -1,0 +1,11 @@
+package com.example.dms_android.feature.auth.login
+
+sealed class SignInViewEffect {
+    object SignInSuccess : SignInEvent()
+    object BadRequestException : SignInEvent()
+    object UnAuthorizedException : SignInEvent()
+    object NotFoundException : SignInEvent()
+    object TooManyRequestException : SignInEvent()
+    object InternalServerException : SignInEvent()
+    object UnKnownException : SignInEvent()
+}
